små ændringer

+++ src/InheritanceComputer/Laptop.java
@@ -5,7 +5,7 @@ package InheritanceComputer;
         //1 additional property
         private String laptopColor;
 
-        public Laptop(String brandName, double GB, boolean OLEDScreen) {
+        public Laptop(String brandName, int GB, boolean OLEDScreen) {
             super(brandName, GB, OLEDScreen);
         }
